CSPII-11513: move some supporting functions and "ignore filters" at module level, to support both local and remote clients
