+++ nuxeo-drive-client/nxdrive/client/local_client.py
@@ -4,14 +4,13 @@ import unicodedata
 from datetime import datetime
 import hashlib
 import os
-import sys
 import shutil
 import re
 import tempfile
 from functools import wraps
-from collections import Iterable
-from functools import partial
 from nxdrive.client.common import BaseClient, UNACCESSIBLE_HASH
+from nxdrive.client.common import register, add_ignore_filter, remove_ignore_filter, \
+    ignore_prefixes, ignore_suffixes, base_is_ignored
 from nxdrive.osi import AbstractOSIntegration
 
 from nxdrive.client.base_automation_client import DOWNLOAD_TMP_FILE_PREFIX
@@ -129,22 +128,6 @@ class FileInfo(object):
         return h.hexdigest()
 
 
-registry = set()
-
-
-def register(*args, **kwargs):
-    def decorator(func):
-        try:
-            apply_to_parent = kwargs['apply_to_parent']
-        except KeyError:
-            apply_to_parent = False
-        func2 = partial(func, *args, **kwargs)
-        func2.apply_to_parent = apply_to_parent
-        registry.add(func2)
-        return func2
-    return decorator
-
-
 class LocalClient(BaseClient):
     """Client API implementation for the local file system"""
 
@@ -159,115 +142,56 @@ class LocalClient(BaseClient):
         # computation if the synchronization thread needs to be suspended
         self.check_suspended = check_suspended
 
-        cls = type(self)
         if ignored_prefixes is not None:
             # remove default prefixes and register with these new ones
-            fname = cls.ignore_prefixes.func.func_name
-            func = cls._remove_ignore_filter(fname)
+            fname = ignore_prefixes.func.func_name
+            func = remove_ignore_filter(fname)
             # reregister the "ignore_prefixes filter with the new prefixes
-            self._add_ignore_filter(fname, prefixes=ignored_prefixes, apply_to_parent=func.apply_to_parent)
+            add_ignore_filter(func, prefixes=ignored_prefixes, apply_to_parent=func.apply_to_parent)
 
         if ignored_suffixes is not None:
             # remove default suffixes and register with these new ones
-            fname = cls.ignore_suffixes.func.func_name
-            func = cls._remove_ignore_filter(fname)
+            fname = ignore_suffixes.func.func_name
+            func = remove_ignore_filter(fname)
             # reregister the "ignore_suffixes filter with the new suffixes
-            self._add_ignore_filter(fname, suffixes=ignored_suffixes, apply_to_parent=func.apply_to_parent)
+            add_ignore_filter(func, suffixes=ignored_suffixes, apply_to_parent=func.apply_to_parent)
 
         while len(base_folder) > 1 and base_folder.endswith(os.path.sep):
             base_folder = base_folder[:-1]
         self.base_folder = base_folder
         self._digest_func = digest_func
 
-    @staticmethod
-    def _remove_ignore_filter(func_name):
-        partial_func = None
-        for f in registry:
-            if f.func.func_name == func_name:
-                partial_func = f
-                break
-        if partial_func:
-            registry.remove(partial_func)
-        return partial_func
-
-    def _add_ignore_filter(self, func_name, **kwargs):
-        partial_func = getattr(self, func_name, None)
-        if partial_func:
-            register(**kwargs)(partial_func.func)
-
     @register()
     def ignore_non_existant(self, parent, name, **kwargs):
+        # 'self' here represents the context where the function is called and is used to facilitate its work; use
+        # 'duck-typing' to check for the right context.
+        # skip if called with a context which cannot compute the absolute path, e.g. missing,
+        # not derived from LocalClient, etc.
+        if self is None or not hasattr(self, '_abspath'):
+            return False
         path = self._abspath(os.path.join(parent, name))
         return not os.path.exists(path)
 
     @register()
     def ignore_guest_folder(self, parent, name, **kwargs):
+        # 'self' here represents the context where the function is called and is used to facilitate its work; use
+        # 'duck-typing' to check for the right context.
+        # skip if called with a context which cannot compute the absolute path, e.g. missing,
+        # not derived from LocalClient, etc.
+        if self is None or not hasattr(self, '_abspath'):
+            return False
         path = self._abspath(os.path.join(parent, name))
         return os.path.exists(path) and os.path.isdir(path) and name == 'Guest Folder'
-
-    @register(prefixes=(
-            '.',  # hidden Unix files
-            '~$',  # Windows lock files
-            'Thumbs.db',  # Thumbnails files
-            'Icon\r',  # Mac Icon
-            'desktop.ini',  # Icon for windows
-    ), apply_to_parent=True
-    )
-    def ignore_prefixes(self, parent, name, **kwargs):
-        prefixes = kwargs['prefixes']
-        if prefixes is None:
-            return False
-        if isinstance(prefixes, Iterable):
-            for prefix in prefixes:
-                if name.startswith(prefix):
-                    return True
-        elif isinstance(prefixes, str):
-            if name.startswith(prefixes):
-                return True
-        else:
-            return False
-
-    @register(suffixes=(
-            '~',  # editor buffers
-            '.swp',  # vim swap files
-            '.lock',  # some process use file locks
-            '.LOCK',  # other locks
-            '.part', '.crdownload', '.partial',  # partially downloaded files by browsers
-    ))
-    def ignore_suffixes(self, parent, name, **kwargs):
-        suffixes = kwargs['suffixes']
-        if suffixes is None:
-            return False
-        if isinstance(suffixes, Iterable):
-            for suffix in suffixes:
-                if name.endswith(suffix):
-                    return True
-        elif isinstance(suffixes, str):
-            if name.endswith(suffixes):
-                return True
-        else:
-            return False
-
-    @register(prefixes=('~', '#',), suffixes=('.tmp', '#',))
-    def ignore_prefix_and_suffix(self, parent, name, **kwargs):
-        prefixes = kwargs['prefixes']
-        suffixes = kwargs['suffixes']
-        if prefixes is None or suffixes is None:
-            return False
-        if len(name) < 2:
-            return False
-        if isinstance(prefixes, str) and isinstance(suffixes, str):
-            return name.startswith(prefixes) and name.endswith(suffixes)
-
-        if isinstance(prefixes, Iterable) and isinstance(suffixes, Iterable):
-            for prefix, suffix in zip(prefixes, suffixes):
-                if name.startswith(prefix) and name.endswith(suffix):
-                    return True
-        return False
 
     @register(apply_to_parent=True)
     def ignore_hidden(self, parent, name, **kwargs):
         if AbstractOSIntegration.is_windows():
+            # 'self' here represents the context where the function is called and is used to facilitate its work; use
+            # 'duck-typing' to check for the right context.
+            # skip if called with a context which cannot compute the absolute path, e.g. missing,
+            # not derived from LocalClient, etc.
+            if self is None or not hasattr(self, '_abspath') or not hasattr(self, 'get_children_ref'):
+                return False
             # NXDRIVE 465
             ref = self.get_children_ref(parent, name)
             path = self._abspath(ref)
@@ -617,24 +541,15 @@ class LocalClient(BaseClient):
 
     def manipulate_guest_filter(self, ignore_guest):
         if ignore_guest:
-            func = None
-            for f in registry:
-                if f.func.func_name == "ignore_guest_folder":
-                    func = f
-                    break
-            if func:
-                registry.remove(func)
+            remove_ignore_filter("ignore_guest_folder")
         else:
-            registry.add(self.ignore_guest_folder)
+            func = getattr(self, "ignore_guest_folder", None)
+            if func:
+                add_ignore_filter(func)
 
     def is_ignored(self, parent_ref, file_name, ignore_guest=False):
         self.manipulate_guest_filter(ignore_guest)
-        grandparent_ref, dir_name = os.path.split(parent_ref)
-        if dir_name:
-            ignore = any(f(grandparent_ref, dir_name) for f in registry if f.apply_to_parent)
-        else:
-            ignore = False
-        return ignore or any(f(self, parent_ref, file_name) for f in registry)
+        return base_is_ignored(self, parent_ref, file_name)
 
     def get_children_ref(self, parent_ref, name):
         if parent_ref == u'/':
